레디스 클라이언트 로직 모듈화
