+++ common/utils/redisClient.js
@@ -1,0 +1,19 @@
+import dotenv from 'dotenv';
+import { createClient } from 'redis';
+
+dotenv.config();
+
+const redisClient = createClient({
+  url: `redis://${process.env.REDIS_URL}:${process.env.REDIS_PORT}`,
+  password: process.env.REDIS_PASSWORD,
+});
+
+redisClient.connect()
+  .then(() => {
+    console.log('Connected to Redis');
+  })
+  .catch((err) => {
+    console.error('Error connecting to Redis:', err);
+  });
+
+export default redisClient;
